Remove API URL check in env-config
The logic for checking if API URL was loaded was removed from env-config.ts. The app should not stop execution if API URL is not found as we are now catering for situations where the app could run without calling backend services. The error handling for missing environment variables should be carried out at a higher level of the application.